+++ src/main/java/gov/antt/sifama/repositories/FotoRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FotoRepo extends JpaRepository<Foto, Long> {
+public interface FotoRepo extends JpaRepository<Foto, Integer> {
 }
