+++ src/main/java/gov/antt/sifama/repositories/LocalRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LocalRepo extends JpaRepository<Local, Long> {
+public interface LocalRepo extends JpaRepository<Local, Integer> {
 }
