+++ src/main/java/gov/antt/sifama/repositories/TroRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TroRepo extends JpaRepository<Tro, Long> {
+public interface TroRepo extends JpaRepository<Tro, Integer> {
 }
